# checked_ilog: improve performance