Addresses #115874.

(This PR replicates the original #115875, which I accidentally closed by deleting my forked repository...)