@tspiteri 2024-04-19
I ran the benchmarks included in the fourth commit. Basically, for "before" I cherry picked just the fourth commit onto the master branch, and for the "after" I rebased the whole PR onto the master. Results below:

```
Before:

    num::int_log::u128_log_geometric        82.91µs/iter   +/- 1.20µs
    num::int_log::u128_log_random            2.10ms/iter   +/- 2.74µs
    num::int_log::u128_log_random_small      1.42ms/iter  +/- 14.30µs
    num::int_log::u16_log_geometric        740.00ns/iter   +/- 0.00ns
    num::int_log::u16_log_random           634.11µs/iter  +/- 10.48µs
    num::int_log::u16_log_random_small     502.78µs/iter   +/- 3.06µs
    num::int_log::u32_log_geometric          4.29µs/iter  +/- 11.00ns
    num::int_log::u32_log_random           757.75µs/iter   +/- 2.19µs
    num::int_log::u32_log_random_small     203.69µs/iter   +/- 2.36µs
    num::int_log::u64_log_geometric         26.20µs/iter +/- 124.00ns
    num::int_log::u64_log_random             1.09ms/iter   +/- 3.65µs
    num::int_log::u64_log_random_small     251.95µs/iter   +/- 3.23µs
    num::int_log::u8_log_geometric         180.00ns/iter   +/- 6.00ns
    num::int_log::u8_log_random            192.59µs/iter   +/- 2.61µs
    num::int_log::u8_log_random_small      192.83µs/iter   +/- 2.88µs

After:

    num::int_log::u128_log_geometric        26.60µs/iter  +/- 65.00ns
    num::int_log::u128_log_random            1.83ms/iter  +/- 11.76µs
    num::int_log::u128_log_random_small    936.31µs/iter   +/- 3.43µs
    num::int_log::u16_log_geometric        296.00ns/iter  +/- 17.00ns
    num::int_log::u16_log_random           466.24µs/iter   +/- 2.04µs
    num::int_log::u16_log_random_small     407.55µs/iter   +/- 9.85µs
    num::int_log::u32_log_geometric        812.00ns/iter   +/- 6.00ns
    num::int_log::u32_log_random           451.27µs/iter   +/- 2.17µs
    num::int_log::u32_log_random_small     468.65µs/iter   +/- 2.36µs
    num::int_log::u64_log_geometric          4.99µs/iter  +/- 30.00ns
    num::int_log::u64_log_random           593.27µs/iter   +/- 1.83µs
    num::int_log::u64_log_random_small     179.98µs/iter   +/- 3.62µs
    num::int_log::u8_log_geometric         129.00ns/iter   +/- 3.00ns
    num::int_log::u8_log_random            475.02µs/iter   +/- 1.58µs
    num::int_log::u8_log_random_small      474.46µs/iter   +/- 1.39µs
```