@rust-log-analyzer 2024-04-22
The job **`x86_64-gnu-aux`** failed! Check out the build log: [(web)](https://github.com/rust-lang-ci/rust/actions/runs/8776442070/job/24079949306) [(plain)](https://github.com/rust-lang-ci/rust/commit/981835df0f70fcaf27988a5cddf8cb7169948e12/checks/24079949306/logs)

<details><summary><i>Click to see the possible cause of the failure (guessed by this bot)</i></summary>

```plain

```

</details>